Move AppState to separate module
