+++ trustchain-http/src/lib.rs
@@ -9,6 +9,7 @@ pub mod handlers;
 pub mod issuer;
 pub mod qrcode;
 pub mod resolver;
+pub mod state;
 pub mod vc;
 pub mod verifier;
